+++ packages/linter/src/index.ts
@@ -34,6 +34,7 @@ import { GoogleFontPreconnect } from "./rules/GoogleFontPreconnect";
 import { BoilerplateIsRemoved } from "./rules/BoilerplateIsRemoved";
 import { AmpImgUsesSrcSet } from "./rules/AmpImgUsesSrcSet";
 import { ViewportDisablesTapDelay } from "./rules/ViewportDisablesTapDelay";
+import { IsUsingLatestComponentVersion } from "./rules/IsUsingLatestComponentVersion";
 import { RuleConstructor } from "./rule";
 import { isArray } from "util";
 import * as cheerio from "cheerio";
@@ -142,17 +143,15 @@ function testsForMode(type: LintMode) {
       IsValid,
       RuntimeIsPreloaded,
       BlockingExtensionsPreloaded,
-      FontsArePreloaded,
-      FastGoogleFontsDisplay,
       NoIconFontIsUsed,
       GoogleFontPreconnect,
       IsTransformedAmp,
       BoilerplateIsRemoved,
       ModuleRuntimeUsed,
-      HeroImageIsDefined,
       AmpImgUsesSrcSet,
       AmpImgPlaceholderIsDataUri,
       ViewportDisablesTapDelay,
+      IsUsingLatestComponentVersion,
     ])
   );
   return tests.get(type) || [];